Use AWS credentials stored in a Secret
* Objective

This commit objective is to use the AWS credentials stored in a Secret
to perform AWS authentication.

* Why

While releasing the controller, we do need to make available the
credentials for the reconciler to operate, a Secret looks like the ideal
place initially.

* How

This commit updates the AWS authentication to prevent "defaults" and use
a static credentials authentication.